requirements.txt
anyio==3.6.1
certifi==2022.6.15
click==8.1.3
fastapi==0.79.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pydantic==1.9.1
python-dotenv==0.20.0
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
